etl/requirements.txt
pyyaml
psycopg2-binary
python-docx
